packages/sdk/pyproject.toml
[project]
name = "handoff-sdk"
version = "0.1.0"
description = "Python SDK for the Handoff agent-to-agent protocol"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.26.0",
    "websockets>=12.0",
    "PyNaCl>=1.5.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
